requirements.txt
tensorflow==1.13.1
numpy==1.16.2
Django==2.2
whitenoise==4.1
Pillow==6.0.0